+++ app/assets/javascripts/application.js
@@ -22,10 +22,10 @@
 //= require es5-shim
 //= require underscore
 //= require metisMenu
+//= require jquery.countdown.min.js
 //= require_tree ./extensions
 //= require components/react_ujs
 //= require_tree ./other
-//= require jquery.countdown
 
 // Load all browserify modules below.