+++ app/assets/javascripts/other/dashboard.js
@@ -1,15 +1,13 @@
 $(document).ready(function () {
-  $("#unresolved_recommendations_notification_modal").modal()
+  $("#unresolved_recommendations_notification_modal").modal();
 
+     $('[data-countdown]').each(function() {
+           var $this = $(this), finalDate = $(this).data('countdown');
+           $this.countdown(finalDate, function(event) {
+                $this.html(event.strftime('%D Days, %H Hours, %M Minutes'));
+               });
+         });
 });
 
 
 
- jQuery(document).ready(function($) {
-    $("#countdown")
-        .countdown("2019/01/01", function(event) {
-            $(this).text(
-                event.strftime('%D days %H:%M:%S')
-            );
-        });
-})