+++ app/controllers/application_controller.rb
@@ -15,7 +15,7 @@ class ApplicationController < ActionController::Base
       ).order("main DESC")
       @travel_plans  = current_account.travel_plans.includes_destinations
       @unresolved_recommendations = current_account.card_recommendations.unresolved
-      @recommendation_expiration = current_account.recommendation_expiry
+      @recommendation_expiration = current_account.recommendations_expire_at
       render "accounts/dashboard"
     else
       redirect_to new_account_session_path